+++ android/app/src/main/kotlin/deckers/thibault/aves/channel/calls/DeviceHandler.kt
@@ -32,10 +32,6 @@ class DeviceHandler(private val context: Context) : MethodCallHandler {
                 "canPinShortcut" to ShortcutManagerCompat.isRequestPinShortcutSupported(context),
                 "canPrint" to (sdkInt >= Build.VERSION_CODES.KITKAT),
                 "canRenderFlagEmojis" to (sdkInt >= Build.VERSION_CODES.LOLLIPOP),
-                // as of google_maps_flutter v2.1.1, minSDK is 20 because of default PlatformView usage,
-                // but using hybrid composition would make it usable on API 19 too,
-                // cf https://github.com/flutter/flutter/issues/23728
-                "canRenderGoogleMaps" to (sdkInt >= Build.VERSION_CODES.KITKAT_WATCH),
                 "showPinShortcutFeedback" to (sdkInt >= Build.VERSION_CODES.O),
                 "supportEdgeToEdgeUIMode" to (sdkInt >= Build.VERSION_CODES.Q),
             )