+++ lib/model/availability.dart
@@ -1,5 +1,6 @@
-import 'package:aves/model/device.dart';
-import 'package:aves_services_platform/aves_services_platform.dart';
+import 'package:aves/model/settings/enums/map_style.dart';
+import 'package:aves/services/common/services.dart';
+import 'package:aves_map/aves_map.dart';
 import 'package:connectivity_plus/connectivity_plus.dart';
 import 'package:flutter/foundation.dart';
 
@@ -10,7 +11,7 @@ abstract class AvesAvailability {
 
   Future<bool> get canLocatePlaces;
 
-  Future<bool> get canUseDeviceMaps;
+  List<EntryMapStyle> get mapStyles;
 }
 
 class LiveAvesAvailability implements AvesAvailability {
@@ -42,11 +43,11 @@ class LiveAvesAvailability implements AvesAvailability {
   // local geocoding with `geocoder` seems to require Google Play Services
   // what about devices with Huawei Mobile Services?
   @override
-  Future<bool> get canLocatePlaces => Future.wait<bool>([
-        isConnected,
-        PlatformMobileServices().isServiceAvailable(),
-      ]).then((results) => results.every((result) => result));
+  Future<bool> get canLocatePlaces async => mobileServices.isServiceAvailable && await isConnected;
 
   @override
-  Future<bool> get canUseDeviceMaps async => device.canRenderGoogleMaps && await PlatformMobileServices().isServiceAvailable();
+  List<EntryMapStyle> get mapStyles => [
+        ...mobileServices.mapStyles,
+        ...EntryMapStyle.values.where((v) => !v.needMobileService),
+      ];
 }