+++ CHANGELOG.md
@@ -17,6 +17,7 @@ All notable changes to this project will be documented in this file.
 
 - upgraded Flutter to stable v3.0.1
 - stretching overscroll effect
+- disabled Google Maps layer on Android Lollipop
 
 ### Fixed